get env var tests
