+++ packages/utils/src/utils/__init__.py
@@ -1,17 +1,3 @@
-def get_env_variable(name: str) -> str:
-    """
-    Grabs a variable by name from the environment. Throws an error if the variable is not present.
+from .utils import get_env_variable
 
-    Args:
-        name (str): Name of the environment variable
-
-    Raises:
-        OSError: raised if environment variable is not present
-
-    Returns:
-        str: Name of the environment variable
-    """
-    value = os.getenv(name)
-    if value is None:
-        raise OSError(f"Missing environment variable: {name}")
-    return value
+__all__ = ["get_env_variable"]
