+++ packages/utils/tests/test_get_env_var.py
@@ -1,16 +1,20 @@
+import pytest
+
+from utils import get_env_variable
 
 
 class TestGetEnvVariable:
     def test_get_env_variable(self, monkeypatch):
         """Test require env."""
         monkeypatch.setenv("TEST_ENV_VAR", "test_value")
-        value = lambda_handler.require_env("TEST_ENV_VAR")
+        value = get_env_variable("TEST_ENV_VAR")
         assert value == "test_value"
 
     def test_require_env_not_set(self, monkeypatch):
         """Test require env not set."""
+        nonexistent_env_var = "NONEXISTENT_ENV_VAR"
         with pytest.raises(
-            ValueError,
-            match=r"NONEXISTENT_ENV_VAR not set as an environment variable\.",
+            OSError,
+            match=f"Missing environment variable: {nonexistent_env_var}",
         ):
-            lambda_handler.require_env("NONEXISTENT_ENV_VAR")
+            get_env_variable(nonexistent_env_var)
